feat: noReportChatDeletion

+++ handlers/middlewares/reportHandled.ts
@@ -1,19 +1,28 @@
-import { CallbackQuery, Update } from "telegraf/types";
+import { CallbackQuery, Message, Update } from "telegraf/types";
+import { config } from "../../utils/config";
 import type { ExtendedContext } from "../../typings/context";
 
-export = (ctx: ExtendedContext<Update.CallbackQueryUpdate<CallbackQuery.DataQuery>>) => {
+export = async (
+	ctx: ExtendedContext<Update.CallbackQueryUpdate<CallbackQuery.DataQuery>>,
+) => {
 	if (ctx.from?.status !== "admin") {
 		return ctx.answerCbQuery("✋ Not permitted!", { cache_time: 600 });
 	}
 
-	const [, chatId, msgId] = ctx.match!;
+	// @ts-expect-error ctx.match is available but not registered in this callback type
+	const [, chatId, msgId] = (ctx.match as string[])!;
 
+	// delete the report in the actual chat
+	await ctx.telegram.deleteMessage(+chatId, +msgId);
+
+	if (config.chats?.noReportChatDeletion) return null;
 	return Promise.all([
 		// delete the report in the report chat
 		ctx.deleteMessage(),
 		// delete the forwarded contextual message in report chat
-		ctx.deleteMessage(ctx.callbackQuery.message?.reply_to_message?.message_id),
-		// delete the report in the actual chat
-		ctx.telegram.deleteMessage(+chatId, +msgId),
+		ctx.deleteMessage(
+			(ctx.callbackQuery.message as Message.TextMessage)?.reply_to_message
+				?.message_id,
+		),
 	]);
 };
